fix: load around is now on base 1/16 like expected

+++ src/world/chunk_manager.hpp
@@ -18,7 +18,7 @@ class ChunkManager
 public:
 	ChunkManager();
 
-	void loadAround(const glm::ivec3& pos);
+	void loadAround(const glm::vec3& pos);
 	void draw(glm::mat4 matrix) const;
 
 	void            setMaterial(const Material& mat);
